+++ client/package.json
@@ -28,6 +28,7 @@
     "vuex": "^3.4.0"
   },
   "devDependencies": {
+    "@types/node": "^14.14.12",
     "@typescript-eslint/eslint-plugin": "^2.33.0",
     "@typescript-eslint/parser": "^2.33.0",
     "@vue/cli-plugin-babel": "~4.5.0",
