+++ client/src/services/auth.service.ts
@@ -1,10 +1,10 @@
 import axios from 'axios';
 import authHeader from './auth-header';
-
-const API_URL = 'http://localhost:4000/';
+import { API_URL } from './socket-io.service';
 
 class AuthService {
   login(user:any) {
+    console.log(API_URL + 'users/authenticate')
     return axios
       .post(API_URL + 'users/authenticate', {
         username: user.username,
