+++ client/src/types.d.ts
@@ -1,4 +1,3 @@
-import { colors } from "vuetify/lib";
 
 declare module 'v-drag';
 declare module 'vue-accessible-color-picker' {
